requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
chromadb>=0.4.0
pandas>=2.0.0
pydantic>=2.0.0
phonenumbers>=8.13.0
sentence-transformers>=2.2.0
pypdf>=3.15.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
